refs #29887 cleaned up canvas view
